[core] Rename the ConstraintSetBase template to ConstraintSetTpl, and the context typedef "ConstraintSetBase" to "ConstraintSet"
+ perform similar op to the related header, add header deprecation for constraint-base.hpp
+ add deprecated typedefs for the previous behaviour
