+++ CHANGELOG.md
@@ -16,6 +16,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 * Increase minimum version of eigenpy to 3.8.0
 * [python] Header `proxsuite-nlp/python/deprecation-policy.hpp` is now deprecated and simply using-declares the templates from eigenpy's corresponding header (introduced in [3.6.0](https://github.com/stack-of-tasks/eigenpy/releases/tag/v3.6.0))
+* Rename the `ConstraintSetBase` template class to `ConstraintSetTpl`, add a deprecated typedef for compatibility
+* Rename header `constraint-base.hpp` (and related files) to `constraint-set.hpp`, change former to a deprecated header `constraint-base.hpp`
 
 ## [0.8.0] - 2024-09-16
