+++ CMakeLists.txt
@@ -184,7 +184,7 @@ if(ENABLE_TEMPLATE_INSTANTIATION)
   set(LIB_TEMPLATE_SOURCES
       ${PROJECT_SOURCE_DIR}/src/block-ldlt.cpp
       ${PROJECT_SOURCE_DIR}/src/constraints.cpp
-      ${PROJECT_SOURCE_DIR}/src/constraint-base.cpp
+      ${PROJECT_SOURCE_DIR}/src/constraint-set.cpp
       ${PROJECT_SOURCE_DIR}/src/constraint-set-product.cpp
       ${PROJECT_SOURCE_DIR}/src/manifold-base.cpp
       ${PROJECT_SOURCE_DIR}/src/cartesian-product.cpp
