+++ include/proxsuite-nlp/context.hpp
@@ -17,7 +17,8 @@ using Results = ResultsTpl<Scalar>;
 using Workspace = WorkspaceTpl<Scalar>;
 using Cost = CostFunctionBaseTpl<Scalar>;
 using Constraint = ConstraintObjectTpl<Scalar>;
-using ConstraintSet = ConstraintSetBase<Scalar>;
+using ConstraintSet = ConstraintSetTpl<Scalar>;
+using ConstraintBase PROXSUITE_NLP_DEPRECATED = ConstraintSetTpl<Scalar>;
 using Function = BaseFunctionTpl<Scalar>;
 using C1Function = C1FunctionTpl<Scalar>;
 using C2Function = C2FunctionTpl<Scalar>;
