+++ src/constraint-set.cpp
@@ -1,9 +1,9 @@
-#include "proxsuite-nlp/constraint-base.hpp"
+#include "proxsuite-nlp/constraint-set.hpp"
 
 namespace proxsuite::nlp {
 
 template struct PROXSUITE_NLP_EXPLICIT_INSTANTIATION_DEFINITION_DLLAPI
-    ConstraintSetBase<context::Scalar>;
+    ConstraintSetTpl<context::Scalar>;
 template struct PROXSUITE_NLP_EXPLICIT_INSTANTIATION_DEFINITION_DLLAPI
     ConstraintObjectTpl<context::Scalar>;
